+++ common/realdos.cpp
@@ -387,7 +387,7 @@ void show_speedstring(
         else
         {
             driver_put_string(speedrow, 16, C_CHOICE_SP_INSTR, g_speed_prompt.c_str());
-            j = g_speed_prompt.length();
+            j = static_cast<int>(g_speed_prompt.length());
         }
         std::strcpy(buf, speedstring);
         int i = (int) std::strlen(buf);
